pyproject.toml
[tool.poetry]
name = "tv-ai-robot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tv_ai_robot"}]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
ipython = "^8.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
